Update train.yml
When making use of the cml-runner it seems that the repository is cloned to the docker container overwriting anything that has been copied to that folder structure such as dat. Since all data is in `.gitignore` it will not clone it. Therefore we need to add to the train.yaml workflow that the data is copied from a data-folder in the docker container.